The entire project
So in this project it was a return to coding and the only code i had a chance to write as a project for my employeer. an issue arose when a program stopped outputting IIF files for quick books and instead moved to a more normal CSV file type. the import process of a CSV was difficult for the employee at the station to do. I was tasked with either finding a program that would do the work and purchase it or making it myself.

I chose Python for its ease of implementation and its strong data processing abilities. The project is mostly just formatting the information but it was specifically made so that all the user had to do was make sure the files were saved in the right spots. Run a .exe. and then import the created file.